order: 描き順を i32 に(3 万 2767 で同点になっていた)。ブロックの REACH で近くの物の升目を広げる。夜の自走の報告

+++ motolii/crates/motolii-doc/src/core/frame.rs
@@ -208,7 +208,7 @@ pub struct CompSpec {
 pub struct LayerPlacement {
     pub transform: glam::Affine2,
     pub world_transform: Option<glam::Affine3A>,
-    pub order: i16,
+    pub order: i32,
     pub opacity: f32,
     pub z: f32,
     pub rotation_x: f32,
